fix: prefer native File over polyfill (in elector)
